+++ src/constants/defaultValue.ts
@@ -2,11 +2,10 @@ import { get } from "lodash";
 import { devConfig } from "~/config";
 import { getOptions } from "~/utils/helpers";
 
-// export const BASE_URL = 'config.REACT_APP_BASE_URL';
-export const BASE_URL = '';
+export const BASE_URL = 'config.REACT_APP_BASE_URL';  
+// export const BASE_URL = 'https://pharma-dashboard.congtyso.com';
 export const DEFAULT_UPLOAD_ACTION = `${BASE_URL}/api/v1/file`;
 export const MAX_UPLOAD_FILE_SIZE_IN_MB = 2;
-
 
 export const DEFAULT_BRANCH_ID = 99999;
 export const STATUS = {