+++ src/modules/salesGroup/components/AssignTeamLead.tsx
@@ -5,13 +5,15 @@ import { debounce, get } from "lodash";
 import React, { useCallback, useMemo, useState } from "react";
 import AvatarShortOrName from "~/components/common/AvatarShortOrName";
 import { EMPLOYEE_LEVEL_VI } from "~/modules/employee/constants";
-import { getShortName } from "~/utils/helpers";
 import { useGetListTeamLeadSalesGroups } from "../salesGroup.hook";
 import useSalesGroupStore from "../salesGroupContext";
+
 type propsType = {
   _id?: string;
   teamLead: any;
 };
+const CLONE_GROUP_TYPE_BENEFIT_EMPLOYEE_LEVEL_VI: any = EMPLOYEE_LEVEL_VI;
+
 export default function AssignTeamLead({
   _id,
   teamLead,
@@ -19,7 +21,7 @@ export default function AssignTeamLead({
   const [open, setOpen] = useState(false);
   const { isSubmitLoading, updateSalesGroup } = useSalesGroupStore();
   const [keyword,setKeyword] = useState("");
-  const query = useMemo(() => (open ? { keyword } : null), [keyword, open]);
+  const query = useMemo(() => (open ? { keyword, salesGroupId : _id } : null), [keyword, open]);
   const [data, isLoading] = useGetListTeamLeadSalesGroups(query);
   const hide = useCallback(() => {
     setOpen(false);
@@ -93,7 +95,7 @@ export default function AssignTeamLead({
           open={open}
           onOpenChange={handleOpenChange}
         >
-          <Tooltip title={teamLead ? "Thay đổi trưởng nhóm" : "Thêm trưởng nhóm"}>
+          <Tooltip title={teamLead ? `Thay đổi ${CLONE_GROUP_TYPE_BENEFIT_EMPLOYEE_LEVEL_VI[get(teamLead, "employee.employeeLevel", "")]}` : `Thêm quản lý`}>
           <Button
             icon={
               teamLead ? (
@@ -106,8 +108,8 @@ export default function AssignTeamLead({
           </Tooltip>
         </Popover>
         {teamLead ? (
-          <Popconfirm title="Xác nhận gỡ trưởng nhóm" onConfirm={onRemove}>
-          <Tooltip title="Gỡ trưởng nhóm">
+          <Popconfirm title= {`Xác nhận gỡ ${CLONE_GROUP_TYPE_BENEFIT_EMPLOYEE_LEVEL_VI[get(teamLead, "employee.employeeLevel", "")]}`} onConfirm={onRemove}>
+          <Tooltip title={`Gỡ ${CLONE_GROUP_TYPE_BENEFIT_EMPLOYEE_LEVEL_VI[get(teamLead, "employee.employeeLevel", "")]}`}>
           <Button danger icon={<DeleteOutlined />} />
           </Tooltip>
           </Popconfirm>