+++ src/modules/salesGroup/components/Member.tsx
@@ -1,5 +1,5 @@
 import { AntDesignOutlined, UserOutlined } from "@ant-design/icons";
-import { Avatar, Button, Flex, Modal, Popover, Tooltip, Typography } from "antd";
+import { Avatar, Button, ConfigProvider, Flex, Modal, Popover, Tooltip, Typography } from "antd";
 import { get } from "lodash";
 import React, { useMemo, useState } from "react";
 import AvatarShortOrName from "~/components/common/AvatarShortOrName";
@@ -12,12 +12,16 @@ import ListMember from "./ListMember";
 import PopoverCardEmployee from "./PopoverCardEmployee";
 import ListMemberModal from "./ListMemberModal";
 import useSalesGroupStore from "../salesGroupContext";
+import {GROUP_TYPE_BENEFIT_EMPLOYEE_LEVEL_VI } from "~/modules/reportSalary/benefitConfiguration/constants";
+import { EMPLOYEE_LEVEL_VI } from "~/modules/employee/constants";
 type propsType = {
   _id?: string;
   typeArea?: TypeAreaType;
   data: MemberRulesInGroupType[];
   child ? : any[]
 };
+const CLONE_GROUP_TYPE_BENEFIT_EMPLOYEE_LEVEL_VI: any = EMPLOYEE_LEVEL_VI;
+
 export default function Member({ _id, data, typeArea, child }: propsType): React.JSX.Element {
   const [isOpen, setIsOpen] = useState(false);
   const { canUpdate } = useSalesGroupStore();
@@ -37,24 +41,36 @@ export default function Member({ _id, data, typeArea, child }: propsType): React
   const onClose = () => {
     setIsOpen(false);
   };
-
   return (
     <>
       <Flex vertical gap={10}>
         <Flex align={"center"} gap={5}>
-          Trưởng nhóm:{" "}
           {teamLead ? (
-            <PopoverCardEmployee employee={get(teamLead, "employee", "")}>
-          <Typography.Text style={{cursor : 'pointer'}} strong>
-              {get(teamLead, "employee.fullName", "")}
-            </Typography.Text>
+            <>
+              {/* {teamLead?.employee?.employeeLevel === 'ASM' || teamLead?.employee?.employeeLevel === 'LEADER'
+                ?  <Typography.Text strong>{CLONE_GROUP_TYPE_BENEFIT_EMPLOYEE_LEVEL_VI[get(teamLead, "employee.employeeLevel", "")]}:</Typography.Text>
+                : <></>
+              } */}
+              <PopoverCardEmployee employee={get(teamLead, "employee", "")}>
+                <ConfigProvider
+                  theme={{
+                    token: {
+                      fontSizeHeading5: 14
+                    },
+                  }}
+                >
+                  <Typography.Title style={{cursor : 'pointer', width: '176px'}} ellipsis level={5}>
+                    {get(teamLead, "employee.fullName", "")}
+                  </Typography.Title>
+                </ConfigProvider>
             </PopoverCardEmployee>
+            </>
           ) : (
-            "(Chưa có)"
+            <span style={{ color: "red" }}>(Chưa có)</span>
           )}{" "}
           {canUpdate && <AssignTeamLead teamLead={teamLead} _id={_id} />}
         </Flex> 
-      {teamLead ?  <Flex align={"center"} gap={5} >
+      {/* {teamLead ?  <Flex align={"center"} gap={5} >
         <Button type="link" onClick={onOpen} style={{whiteSpace : 'nowrap'}}>Thành viên</Button>:
           {member ? (
             <ListMember member={member}/>
@@ -62,7 +78,7 @@ export default function Member({ _id, data, typeArea, child }: propsType): React
             "(Chưa có)"
           )}
           <AssignMember member={member} _id={_id} />
-        </Flex>  : <></>}
+        </Flex>  : <></>} */}
       </Flex>
       <Modal
         title={`Danh sách thành viên`}