+++ src/modules/salesGroup/salesGroup.hook.ts
@@ -183,3 +183,11 @@ export const useResetAction = () => {
       param
     });
   };
+
+  // export const convertData = (data: any) => {
+  //   return [
+  //     ...data.map((item: any) => ({
+        
+  //     }))
+  //   ]
+  // }